cast {{{debugLevel}}} to {{{int}}} to avoid warning; refs #90

+++ tksrc/GMTK_Signals.cc
@@ -87,7 +87,7 @@ void catch_increment(int sig_num)
 void catch_decrement(int sig_num)
 {
   unsigned debugLevel = IM::glbMsgLevel(IM::Inference);
-  if (debugLevel >= debugIncrement)
+  if ((int)debugLevel >= debugIncrement)
     debugLevel -= debugIncrement;
   else
     debugLevel = 0;
